change int to bigint

+++ lib/generators/beta_feature/templates/install.rb
@@ -1,7 +1,7 @@
 class <%= migration_class_name %> < <%= migration_parent %>
   def self.up
     create_table :beta_feature_settings, force: true do |t|
-      t.integer :betable_id, null: false
+      t.bigint :betable_id, null: false
       t.string :betable_type, null: false
       t.string :betas, array: true, default: [], null: false
       t.timestamps
